1.13.2 mapping updates 3

+++ net/minecraft/server/DataConverterItemFrame.java
@@ -4,7 +4,6 @@ import com.mojang.datafixers.DSL;
 import com.mojang.datafixers.Dynamic;
 import com.mojang.datafixers.Typed;
 import com.mojang.datafixers.schemas.Schema;
-import java.util.function.Function;
 
 public class DataConverterItemFrame extends DataConverterNamedEntity {
 
@@ -23,14 +22,14 @@ public class DataConverterItemFrame extends DataConverterNamedEntity {
     private static byte a(byte b0) {
         switch (b0) {
         case 0:
-            return (byte) 3;
+            return 3;
         case 1:
-            return (byte) 4;
+            return 4;
         case 2:
         default:
-            return (byte) 2;
+            return 2;
         case 3:
-            return (byte) 5;
+            return 5;
         }
     }
 }
