+++ net/minecraft/server/NBTTagFloat.java
@@ -24,7 +24,7 @@ public class NBTTagFloat extends NBTNumber {
     }
 
     public byte getTypeId() {
-        return (byte) 5;
+        return 5;
     }
 
     public String toString() {
